Refactored tests for JWT encoding and decoding to separate test files.
